roi shot 명령어 제거
roi shot으로 하던 일을 이제는 웹으로 할 수 있게 되었기 때문에
더이상 쓸모가 없어졌다.

이제는 로이의 기능이 서버를 실행하는 것 밖에 없기 때문에
roi server가 roi가 되었다.

Close: #563
